refactor: extract LinkAddress from LinkWallet and default to linking to an address for unknown types
