+++ src/components/links/index.ts
@@ -1,5 +1,6 @@
 import LinkBlock from "./LinkBlock.vue";
 import LinkTransaction from "./LinkTransaction.vue";
 import LinkWallet from "./LinkWallet.vue";
+import LinkAddress from "./LinkAddress.vue";
 
-export { LinkBlock, LinkTransaction, LinkWallet };
+export { LinkBlock, LinkTransaction, LinkWallet, LinkAddress };
